upgrade formulae using the same compilers as installed them
If a formula needs a non-default compiler to install, then most likely
it will need the same non-default compiler to upgrade. This commit
makes the `upgrade` command use the same compiler to upgrade each
formula as was used to install the formula.

Function parameters are introduced to enable a future commit to
implement a `--cc=` command-line option for `upgrade` to use a
different compiler than that which was used for formula installation.