Strategy: Use GET for checking response head again
Using a `HEAD` request when checking the response head fixes the
check for one cask but breaks 33 existing `HeaderMatch` checks, so
this seems like a poor trade-off at the moment. In the long term,
we will use a `HEAD` request as the default and account for checks
that require a `GET` request in the relevant `livecheck` block but
this isn't possible yet (support is forthcoming).

In the interim time, returning to using `GET` requests by default
seems sensible in light of these broken checks.