caveats: suggest command that preserves existing `PKG_CONFIG_PATH`
Before:

    ❯ brew info curl
    [snip]
    For pkg-config to find curl you may need to set:
      export PKG_CONFIG_PATH="/usr/local/opt/curl/lib/pkgconfig"
    [snip]

After:

    ❯ brew info curl
    [snip]
    For pkg-config to find curl you may need to set:
      export PKG_CONFIG_PATH="/usr/local/opt/curl/lib/pkgconfig${PKG_CONFIG_PATH:+:${PKG_CONFIG_PATH}}"
    [snip]

This means that the suggested command will prepend the suggested path to
`PKG_CONFIG_PATH` if it's already set, or set `PKG_CONFIG_PATH` to the
suggested path if it is not set or is null.

I've only implemented this for Bash-like shells, because I don't know
what the syntax is for other shells.

Closes #15546.