formula: add slimmer api json representation
This is a second go at this after the approach in the earlier PR
ended up not yielding big gains. IMO this is the simplest approach
to reducing the JSON size by omitting the bottle url field that we
don't use anywhere and also removing blank fields at the top-level
of the resulting hash.