dev-cmd/generate-formula-api: have internal json match api data structure
We want the internal JSON used to send the package list over the internet
instead of using the tap Git repo to match the internal data representation
we have for the core formula tap. That means including additional information
like aliases, renames and tap migrations along with storing the formulae
as a hash of name to formula info. This not only reduces extra data
transformation when loading the data but also reduces network calls since
tap migrations are now included in the same file.

The amount of additional data included in the file is very small so this
shouldn't drastically affect the file size either.