Sort printed package list in `brew upgrade`
At the beginning of a `brew upgrade` run, brew prints the list of
packages to be upgraded e.g. as

    ncdu 2.2.2 -> 2.3
    nmap 7.94 -> 7.94_1
    jq 1.6 -> 1.7.1
    sdl2 2.26.5 -> 2.30.0
    libusb 1.0.26 -> 1.0.27
    python@3.11 3.11.3 -> 3.11.7_1
    file-formula 5.44 -> 5.45

The order is apparently arbitrary (I think maybe FS readdir order?)
which makes it difficult to easily eyeball the list to see if some
specific package of interest is included. Instead, sort it.