Clean up test cleanup
Before this change we'd only selectively delete files and directories
in the temporary test directory in between tests. Now we just nuke
all files and directories within that temporary test directory. This
is not only simpler but also allows us to remove a helper and a bunch
of manual cleanup.