bump*: tweak autobump message
We've had multiple issues filed because BrewTestBot isn't bumping
formulae, and they're not able to `brew bump` it.

Let's adjust the messaging to encourage them to open a PR manually
instead.

We should probably also make BrewTestBot ignore open PRs that aren't
version-bump PRs, but I'd prefer to punt on that for now.